Bugfix: Implementing validation on buttons with Formik

+++ src/pages/DeleteStore/index.js
@@ -9,6 +9,7 @@ import Header from "../../util/Header/header";
 import { RegisterButton } from "../../util/Style/global";
 import { DeletePage } from "./style";
 import api from "../../services/requestAPI";
+import RegexCnpj from "../../util/Regex/regexCnpj";
 
 toast.configure();
 
@@ -16,18 +17,27 @@ export default function DeleteStore() {
 	const history = useHistory();
 
 	const handleSubmit = async (values) => {
+		const newCnpj = RegexCnpj(values.cnpj);
+
 		try {
-			await api.delete(`/store/${values.cnpj}`);
+			if (newCnpj.length < 14) {
+				toast.error("CNPJ Inválido", {
+					position: toast.POSITION.TOP_CENTER,
+					autoClose: 3000,
+				});
+			} else {
+				await api.delete(`/store/${values.cnpj}`);
 
-			toast.success("Loja deletada com sucesso!", {
-				position: toast.POSITION.TOP_CENTER,
-				autoClose: 3000,
-			});
+				toast.success("Loja deletada com sucesso!", {
+					position: toast.POSITION.TOP_CENTER,
+					autoClose: 3000,
+				});
 
-			if (sessionStorage.getItem("permission") === "project") {
-				history.push("/homeproject");
-			} else if (sessionStorage.getItem("permission") === "master") {
-				history.push("/homemaster");
+				if (sessionStorage.getItem("permission") === "project") {
+					history.push("/homeproject");
+				} else if (sessionStorage.getItem("permission") === "master") {
+					history.push("/homemaster");
+				}
 			}
 		} catch (erro) {
 			toast.error(erro.response.data.error, {
@@ -54,34 +64,42 @@ export default function DeleteStore() {
 					onSubmit={handleSubmit}
 					validationSchema={validations}
 				>
-					<Form className="delete-container__form">
-						<div className="delete-container__form__group">
-							<ErrorMessage name="cnpj" component="span" />
-							<Field
-								name="cnpj"
-								placeholder="CNPJ:"
-								type="text"
-								className="delete-container__form__cnpj"
-							></Field>
-							<p className="delete-container__form__cnpjIcon"></p>
-						</div>
-						<div className="delete-container__form__buttons">
-							<RegisterButton
-								onClick={() => {
-									if (sessionStorage.getItem("permission") === "project") {
-										history.push("/homeproject");
-									} else if (
-										sessionStorage.getItem("permission") === "master"
-									) {
-										history.push("/homemaster");
-									}
-								}}
-							>
-								Voltar
-							</RegisterButton>
-							<RegisterButton type="submit">Excluir loja</RegisterButton>
-						</div>
-					</Form>
+					{({ dirty, isValid }) => {
+						const submitIsValid = !dirty || !isValid;
+
+						return (
+							<Form className="delete-container__form">
+								<div className="delete-container__form__group">
+									<ErrorMessage name="cnpj" component="span" />
+									<Field
+										name="cnpj"
+										placeholder="CNPJ:"
+										type="text"
+										className="delete-container__form__cnpj"
+									></Field>
+									<p className="delete-container__form__cnpjIcon"></p>
+								</div>
+								<div className="delete-container__form__buttons">
+									<RegisterButton
+										onClick={() => {
+											if (sessionStorage.getItem("permission") === "project") {
+												history.push("/homeproject");
+											} else if (
+												sessionStorage.getItem("permission") === "master"
+											) {
+												history.push("/homemaster");
+											}
+										}}
+									>
+										Voltar
+									</RegisterButton>
+									<RegisterButton type="submit" disabled={submitIsValid}>
+										Excluir loja
+									</RegisterButton>
+								</div>
+							</Form>
+						);
+					}}
 				</Formik>
 			</section>
 			<section className="delete-style"></section>
